Convert the user script to a chrome extension
This is a pretty straightforward port. In 12 years there'd been only a
minor change to the DOM on HN comment pages. I also needed to port the
Greasemonkey get/setValue methods to Chrome's local storage. And create
the rest of the extension boilerplate. But that's about it!

What's fun about this is that Chrome's first stable release was on the
day before this original user script was released. Chrome extensions
weren't released until more than a year later.

So this Chrome extension predates Chrome extensions by about a year.