Implement edit and delete
Two things:

1. There's a lot going on in the PageSpecForm component now and some of
that it inevitably going to need to be refactored out into sub-components.

2. I'm aware that there's a potential for stupid timing bugs in the core
Trakum class, because init and save both happen asynchronously, but
none of the functions in that class take a callback. It only works
correctly because the order that different pieces of the extension are
loaded is favorable. But obviously this is grotesque.