Replace the feature check in MessageRouter with a try/catch
The original idea here was to prevent errors when running the popup in
development mode, but the issue is that sendMessage always exists, it
just has a different signature on web pages than it does for scripts
runnnig in an extension. The original feature check was preventing it
from running correctly when running as a content script. (You get
different levels of access to chrome methods depending on how and where
the code is running.)

The easiest thing to do here is just to catch and log errors.