Refactor/simplify the content script
The content script is going to need to support the different marking and
caching strategies that are forthcoming, but it was already pretty
difficult to work with and those changes would have been difficult to
add without making it even worse.

This refactoring simplifies things quite a bit and clearly sets out the
points at which the cache is retrieved/updated/saved, so that different
strategies can be injected (hopefully) pretty easily.