Inject an element-selector into the page
This is all being done with jQuery, for now; this will probably end up
serving as a proof-of-concept if I can find a reasonble workflow for
dealing with react in a content script. If I can easily mount the react
app on the page, then that's probably what I'll do.