Move some of selector creation into React
Convert the useMousePosition hook into something that can be used to
create a selector for any event and then use this to create and display
selectors on mouseover and clicks.

This functionality can now be moved out of picker.js (and ultimately
everything else in there should be moved into React, as well).